Fix: Failed to parse OpenAI response into ResponseStreamResult: data did not match any variant of untagged enum ResponseStreamResult
for together ai, tool call id was "-1" which was not supported by usize
type.

+++ crates/language_models/src/provider/open_ai.rs
@@ -497,7 +497,7 @@ fn add_message_content_part(
 }
 
 pub struct OpenAiEventMapper {
-    tool_calls_by_index: HashMap<usize, RawToolCall>,
+    tool_calls_by_index: HashMap<i32, RawToolCall>,
 }
 
 impl OpenAiEventMapper {
